Phase 4: Frontend UI + Free Tier audit fixes
- Add playful Cat vs Dog frontend with light/dark mode toggle
- Add drag-and-drop image upload with animated probability bars
- Fix critical bug: apply StandardScaler before ONNX inference
- Add OOD detection for non-cat/dog images (51% threshold)
- Create .dockerignore to exclude data/notebooks/training artifacts
- Update Dockerfile to serve frontend via FastAPI static files
- Add scaler.joblib to repo (required for inference pipeline)
- Update schemas to return per-class probabilities

+++ api/feature_extractor.py
@@ -130,4 +130,4 @@ def process_raw_image(img_bgr):
     ])
     
     features = np.nan_to_num(features, nan=0.0, posinf=0.0, neginf=0.0)
-    return features.reshape(1, -1)
+    return features.reshape(1, -1).astype(np.float32)
